+++ CoffeeMachine/CoffeeMachine/Global.swift
